requirements.txt
pandas
openpyxl
matplotlib
openai
python-dotenv
langfuse
